Resolve comment
