+++ src/ir_util/mod.rs
@@ -1,7 +1,11 @@
 //! utils for working with HIR and MIR
 
 use rustc_hir::definitions::DefPathData;
-use rustc_middle::{query::IntoQueryParam, ty::TyCtxt};
+use rustc_middle::{
+    mir::{Body, TerminatorKind},
+    query::IntoQueryParam,
+    ty::TyCtxt,
+};
 use rustc_span::{Symbol, def_id::DefId};
 
 #[inline]
@@ -32,6 +36,33 @@ pub fn fmt_def_id(
             write!(f, "{}", def_id.index.index())
         }
     })
+}
+
+pub fn body_to_str(body: &Body<'_>) -> String {
+    use std::fmt::Write;
+    let mut s = String::new();
+    writeln!(s, "{:?} {{", body.source.instance.def_id()).unwrap();
+    for (bb, bbd) in body.basic_blocks.iter_enumerated() {
+        writeln!(s, "    {bb:?}:").unwrap();
+        for stmt in &bbd.statements {
+            writeln!(s, "        {stmt:?}").unwrap();
+        }
+        if !matches!(
+            bbd.terminator().kind,
+            TerminatorKind::Return | TerminatorKind::Assert { .. }
+        ) {
+            writeln!(s, "        {:?}", bbd.terminator().kind).unwrap();
+        }
+    }
+    writeln!(s, "}}").unwrap();
+    s
+}
+
+pub fn body_size(body: &Body<'_>) -> usize {
+    body.basic_blocks
+        .iter()
+        .map(|bbd| bbd.statements.len() + 1)
+        .sum()
 }
 
 pub mod ast_to_hir;
